Add filter actions (auto mark read) at category level fix https://github.com/FreshRSS/FreshRSS/issues/3497
