+++ app/Models/Entry.php
@@ -660,39 +660,24 @@ HTML;
 
 	/** @param array<string,bool> $titlesAsRead */
 	public function applyFilterActions(array $titlesAsRead = []): void {
-		if ($this->feed != null) {
-			if (!$this->isRead()) {
-				if ($this->feed->attributes('read_upon_reception') ||
-					($this->feed->attributes('read_upon_reception') === null && FreshRSS_Context::$user_conf->mark_when['reception'])) {
-					$this->_isRead(true);
-					Minz_ExtensionManager::callHook('entry_auto_read', $this, 'upon_reception');
-				}
-				if (!empty($titlesAsRead[$this->title()])) {
-					Minz_Log::debug('Mark title as read: ' . $this->title());
-					$this->_isRead(true);
-					Minz_ExtensionManager::callHook('entry_auto_read', $this, 'same_title_in_feed');
-				}
+		if ($this->feed === null) {
+			return;
+		}
+		if (!$this->isRead()) {
+			if ($this->feed->attributes('read_upon_reception') ||
+				($this->feed->attributes('read_upon_reception') === null && FreshRSS_Context::$user_conf->mark_when['reception'])) {
+				$this->_isRead(true);
+				Minz_ExtensionManager::callHook('entry_auto_read', $this, 'upon_reception');
 			}
-			foreach ($this->feed->filterActions() as $filterAction) {
-				if ($this->matches($filterAction->booleanSearch())) {
-					foreach ($filterAction->actions() as $action) {
-						switch ($action) {
-							case 'read':
-								if (!$this->isRead()) {
-									$this->_isRead(true);
-									Minz_ExtensionManager::callHook('entry_auto_read', $this, 'filter');
-								}
-								break;
-							case 'star':
-								$this->_isFavorite(true);
-								break;
-							case 'label':
-								//TODO: Implement more actions
-								break;
-						}
-					}
-				}
+			if (!empty($titlesAsRead[$this->title()])) {
+				Minz_Log::debug('Mark title as read: ' . $this->title());
+				$this->_isRead(true);
+				Minz_ExtensionManager::callHook('entry_auto_read', $this, 'same_title_in_feed');
 			}
+		}
+		$this->feed->applyFilterActions($this);
+		if ($this->feed->category() !== null) {
+			$this->feed->category()->applyFilterActions($this);
 		}
 	}
